Make general case work for dim = 1.

+++ src/Scene.hpp
@@ -19,7 +19,7 @@ struct ImGuiState {
 	int indx_time = 0;
 	int range_time = 3;
 	ImVec4 offset_pos = ImVec4(0,0,0,0);
-	float offset_time = 0;
+	double offset_time = 0;
 
 	virtual ~ImGuiState() = default;
 };
